+++ src/components/graphs/ExportButtons/ExportButtons.js
@@ -6,7 +6,7 @@ import {
   csvButtonLabel,
   downloadGraphDataLabel, xslxButtonLabel,
 }
-  from '../../guidance-content/info/LabelWithInfoItems';
+  from '../../guidance-content/info/InformationItems';
 
 import styles from './ExportButtons.css';
