+++ src/components/map-controllers/DualMapController/DualMapController.js
@@ -43,7 +43,7 @@ import { hasValidData, currentDataSpec,
   getDatasetId} from '../map-helpers.js';
 
 import styles from '../MapController.css';
-import {mapPanelLabel} from "../../guidance-content/info/LabelWithInfoItems";
+import {mapPanelLabel} from "../../guidance-content/info/InformationItems";
 
 
 // TODO: https://github.com/pacificclimate/climate-explorer-frontend/issues/125
