+++ src/components/map-controllers/SingleMapController/SingleMapController.js
@@ -38,7 +38,7 @@ import { hasValidData, selectRasterPalette,
          updateLayerTime, getTimeParametersPromise, scalarParams} from '../map-helpers.js';
 
 import styles from '../MapController.css';
-import {mapPanelLabel} from "../../guidance-content/info/LabelWithInfoItems";
+import {mapPanelLabel} from "../../guidance-content/info/InformationItems";
 
 
 // TODO: https://github.com/pacificclimate/climate-explorer-frontend/issues/125
